S12-E06

+++ MadPay724.Presentation/Controllers/Site/Admin/PhotosController.cs
@@ -1,4 +1,5 @@
 ﻿
+using System.Linq;
 using System.Security.Claims;
 using System.Threading.Tasks;
 using AutoMapper;
@@ -58,7 +59,6 @@ namespace MadPay724.Presentation.Controllers.Site.Admin
             var uplaodRes = await _uploadService.UploadProfilePic(
                 photoForProfileDto.File,
                 userId,
-                User.FindFirst(ClaimTypes.Name).Value,
                 _env.WebRootPath ,
                 string.Format("{0}://{1}{2}", Request.Scheme, Request.Host.Value, Request.PathBase.Value)
                 );
@@ -66,14 +66,27 @@ namespace MadPay724.Presentation.Controllers.Site.Admin
             if (uplaodRes.Status)
             {
                 photoForProfileDto.Url = uplaodRes.Url;
-                photoForProfileDto.PublicId = uplaodRes.PublicId;
+                if(uplaodRes.LocalUploaded)
+                    photoForProfileDto.PublicId = "1";
+                else
+                    photoForProfileDto.PublicId = uplaodRes.PublicId;
+
 
 
                 var oldphoto = await _db.PhotoRepository.GetAsync(p => p.UserId == userId && p.IsMain);
 
-                if (oldphoto.PublicId != null && oldphoto.PublicId != "0")
+                if (oldphoto.PublicId != null && oldphoto.PublicId != "0" && oldphoto.PublicId != "1")
                 {
                     _uploadService.RemoveFileFromCloudinary(oldphoto.PublicId);
+                }
+                if (oldphoto.PublicId == photoForProfileDto.PublicId && photoForProfileDto.Url.Split('/').Last() != oldphoto.Url.Split('/').Last())
+                {
+                    _uploadService.RemoveFileFromLocal(oldphoto.Url.Split('/').Last(), _env.WebRootPath, "Files\\Pic\\Profile");
+                }
+
+                if (oldphoto.PublicId == "1" && photoForProfileDto.PublicId != "1")
+                {
+                    _uploadService.RemoveFileFromLocal(oldphoto.Url.Split('/').Last(), _env.WebRootPath, "Files\\Pic\\Profile");
                 }
 
                 _mapper.Map(photoForProfileDto, oldphoto);
